add auth

+++ internal/repository/banners.go
@@ -2,14 +2,15 @@ package repository
 
 import (
 	"database/sql"
-	"github.com/lib/pq"
 	"net/http"
 	"time"
 
 	"github.com/jmoiron/sqlx"
 	"github.com/labstack/echo/v4"
+	"github.com/lib/pq"
 
 	"github.com/MisterZurg/avito-tech-internship-2024/api"
+	"github.com/MisterZurg/avito-tech-internship-2024/internal/auth"
 )
 
 type BannersRepository struct {
@@ -29,6 +30,10 @@ type Banner struct {
 // Получение всех баннеров c фильтрацией по фиче и/или тегу
 // (GET /banner)
 func (br *BannersRepository) GetBanner(ctx echo.Context, params api.GetBannerParams) error {
+	if *params.Token != auth.UserToken && *params.Token != auth.AdminToken {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не авторизован")
+	}
+
 	bannerResponse := make([]Banner, 0)
 	// if !(params.FeatureId != nil || params.TagId != nil) {
 	rows, err := br.DB.Queryx(`
@@ -40,7 +45,7 @@ func (br *BannersRepository) GetBanner(ctx echo.Context, params api.GetBannerPar
 				b.created_at,
 				b.updated_at,
 				(SELECT array_agg(tag_id) FROM banners_relation AS br WHERE br.banner_id = b.id) AS tag_ids
-			FROM banners as b
+			FROM banners AS b
 			ORDER BY b.created_at DESC
 			LIMIT $1 OFFSET $2
 			`,
@@ -76,10 +81,18 @@ func (br *BannersRepository) GetBanner(ctx echo.Context, params api.GetBannerPar
 // Создание нового баннера
 // (POST /banner)
 func (br *BannersRepository) PostBanner(ctx echo.Context, params api.PostBannerParams) error {
+	if *params.Token != "" {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не авторизован")
+	}
+
+	if *params.Token != auth.AdminToken {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не имеет доступа")
+	}
+
 	banner := new(api.PostBannerJSONBody)
 
 	if err := ctx.Bind(&banner); err != nil {
-		return ctx.JSON(http.StatusBadRequest, "invalid format for PostBanner")
+		return ctx.JSON(http.StatusInternalServerError, "Внутренняя ошибка сервера")
 	}
 
 	_ = br.DB.QueryRowx(`
@@ -107,6 +120,14 @@ func (br *BannersRepository) PostBanner(ctx echo.Context, params api.PostBannerP
 // Удаление баннера по идентификатору
 // (DELETE /banner/{id})
 func (br *BannersRepository) DeleteBannerId(ctx echo.Context, id int, params api.DeleteBannerIdParams) error {
+	if *params.Token != "" {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не авторизован")
+	}
+
+	if *params.Token != auth.AdminToken {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не имеет доступа")
+	}
+
 	_ = br.DB.QueryRowx(
 		`DELETE from banners WHERE id = $1`,
 		id,
@@ -117,6 +138,14 @@ func (br *BannersRepository) DeleteBannerId(ctx echo.Context, id int, params api
 // Обновление содержимого баннера
 // (PATCH /banner/{id})
 func (br *BannersRepository) PatchBannerId(ctx echo.Context, id int, params api.PatchBannerIdParams) error {
+	if *params.Token != "" {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не авторизован")
+	}
+
+	if *params.Token != auth.AdminToken {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не имеет доступа")
+	}
+
 	row := br.DB.QueryRowx(`
 		SELECT
 			b.id,
@@ -125,8 +154,8 @@ func (br *BannersRepository) PatchBannerId(ctx echo.Context, id int, params api.
 			b.is_active,
 			b.created_at,
 			b.updated_at,
-			(SELECT ARRAY_AGG(tag_id) FROM banners_relation AS br WHERE br.banner_id = b.id) as tag_ids
-		FROM banners as b
+			(SELECT ARRAY_AGG(tag_id) FROM banners_relation AS br WHERE br.banner_id = b.id) AS tag_ids
+		FROM banners AS b
 		WHERE b.id = $1;
 	`,
 		id,
@@ -142,13 +171,13 @@ func (br *BannersRepository) PatchBannerId(ctx echo.Context, id int, params api.
 		(*pq.Int64Array)(&banner.TagIDs),
 	)
 	if err == sql.ErrNoRows {
-		return ctx.JSON(http.StatusNotFound, "")
+		return ctx.JSON(http.StatusNotFound, "Баннер не найден")
 	}
 
 	updBanner := new(api.PostBannerJSONBody)
 
 	if err := ctx.Bind(&updBanner); err != nil {
-		return ctx.JSON(http.StatusBadRequest, "invalid format for PostBanner")
+		return ctx.JSON(http.StatusInternalServerError, "Внутренняя ошибка сервера")
 	}
 
 	_ = br.DB.QueryRowx(`
@@ -164,13 +193,21 @@ func (br *BannersRepository) PatchBannerId(ctx echo.Context, id int, params api.
 		updBanner.Content,
 	)
 
-	return ctx.JSON(http.StatusBadRequest, "")
+	return ctx.JSON(http.StatusBadRequest, "OK")
 
 }
 
 // Получение баннера для пользователя
 // (GET /user_banner)
 func (br *BannersRepository) GetUserBanner(ctx echo.Context, params api.GetUserBannerParams) error {
+	if *params.Token != "" {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не авторизован")
+	}
+
+	if *params.Token != auth.AdminToken {
+		return ctx.JSON(http.StatusUnauthorized, "Пользователь не имеет доступа")
+	}
+
 	row := br.DB.QueryRowx(`
 	WITH find_banner AS (
 		SELECT banner_id, tag_id, feature_id FROM banners_relation WHERE tag_id=$1 AND feature_id=$2
@@ -188,6 +225,9 @@ func (br *BannersRepository) GetUserBanner(ctx echo.Context, params api.GetUserB
 		params.TagId,
 		params.FeatureId,
 	)
+	if row == nil {
+		return ctx.JSON(http.StatusNotFound, "Баннер для не найден")
+	}
 
 	userBanner := new(Banner)
 	err := row.Scan(
@@ -200,7 +240,7 @@ func (br *BannersRepository) GetUserBanner(ctx echo.Context, params api.GetUserB
 		(*pq.Int64Array)(&userBanner.TagIDs),
 	)
 	if err != nil {
-		return ctx.JSON(http.StatusBadRequest, "cannot fill Banner GetUserBanner")
+		return ctx.JSON(http.StatusInternalServerError, "Внутренняя ошибка сервера")
 	}
 	return ctx.JSON(http.StatusOK, userBanner.Content)
 }
